+++ test/plugin/mock/MockAircraftIntentionCodeGenerator.cpp
@@ -1,0 +1,6 @@
+#include "MockAircraftIntentionCodeGenerator.h"
+
+namespace UKControllerPluginTest::IntentionCode {
+    MockAircraftIntentionCodeGenerator::MockAircraftIntentionCodeGenerator() = default;
+    MockAircraftIntentionCodeGenerator::~MockAircraftIntentionCodeGenerator() = default;
+} // namespace UKControllerPluginTest::IntentionCode